Implement linux role.
This role lets the linux role install build tools, clone the IoRegionFd
enabled kernel, and build it. It uses the kernel config for the already
running kernel. It also applies patches to avoid any compiler errors.
Finally it installs the kernel and reboots.

The role is a idempotent as possible for the short amount of time.

The cspell dictionary was also updated with any new words.